+++ praw/const.py
@@ -101,6 +101,7 @@ API_PATH = {
     'modmail_unarchive':      'api/mod/conversations/{id}/unarchive',
     'modmail_unmute':         'api/mod/conversations/{id}/unmute',
     'modmail_unread':         'api/mod/conversations/unread',
+    'modmail_unread_count':   'api/mod/conversations/unread/count',
     'multireddit':            'user/{user}/m/{multi}/',
     'multireddit_api':        'api/multi/user/{user}/m/{multi}/',
     'multireddit_base':       'api/multi/',